pyproject.toml
[tool.poetry]
name = "dgvgkbot"
version = "0.1.0"
description = ""
authors = ["PederHA <peder.andresen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
psutil = "^5.6.7"
httpx = "^0.13.3"
recordclass = "^0.13.2"
ciso8601 = "^2.1.3"
mcstatus = "^2.3.0"
"discord.py" = "^1.3.1"
pyyaml = "^5.4.1"
aiofile = "^3.3.3"
pytz = "^2021.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
